Set TransactionType to MOTO when cvn not available.
Cvn is required for an onsite purchase. It is not required when processing a recurring payment.
It seems reasonable to assume that if not passed, and we are working with a DirectPayment with
a token then it is a recurring payment.

Setting to MOTO or Recurring would have the same effect. The practical difference between Recurring & MOTO is not clear
and there is not an obvious Omnipay standard to follow so sticking with this assumption for
now